+++ psutil/_psutil_windows.c
@@ -201,6 +201,9 @@ psutil_get_nic_addresses() {
  * ============================================================================
  */
 
+// Raised by Process.wait().
+static PyObject *TimeoutExpired;
+static PyObject *TimeoutAbandoned;
 
 static ULONGLONG (*psutil_GetTickCount64)(void) = NULL;
 
@@ -400,20 +403,33 @@ psutil_proc_wait(PyObject *self, PyObject *args) {
     retVal = WaitForSingleObject(hProcess, timeout);
     Py_END_ALLOW_THREADS
 
+    // handle return code
     if (retVal == WAIT_FAILED) {
         CloseHandle(hProcess);
-        return PyErr_SetFromWindowsErr(0);
+        PyErr_SetFromWindowsErr(0);
+        return NULL;
     }
     if (retVal == WAIT_TIMEOUT) {
         CloseHandle(hProcess);
-        return Py_BuildValue("l", WAIT_TIMEOUT);
+        PyErr_SetString(TimeoutExpired,
+                        "WaitForSingleObject() returned WAIT_TIMEOUT");
+        return NULL;
+    }
+    if (retVal == WAIT_ABANDONED) {
+        psutil_debug("WaitForSingleObject() -> WAIT_ABANDONED");
+        CloseHandle(hProcess);
+        PyErr_SetString(TimeoutAbandoned,
+                        "WaitForSingleObject() returned WAIT_ABANDONED");
+        return NULL;
     }
 
+    // WaitForSingleObject() returned WAIT_OBJECT_0. It means the
+    // process is gone so we can get its process exit code. The PID
+    // may still stick around though but we'll handle that from Python.
     if (GetExitCodeProcess(hProcess, &ExitCode) == 0) {
         CloseHandle(hProcess);
         return PyErr_SetFromWindowsErr(GetLastError());
     }
-
     CloseHandle(hProcess);
 
 #if PY_MAJOR_VERSION >= 3
@@ -2847,7 +2863,7 @@ psutil_proc_info(PyObject *self, PyObject *args) {
                 (double)process->UserTime.LowPart * LO_T;
     kernel_time = (double)process->KernelTime.HighPart * HI_T + \
                     (double)process->KernelTime.LowPart * LO_T;
-    
+
     // Convert the LARGE_INTEGER union to a Unix time.
     // It's the best I could find by googling and borrowing code here
     // and there. The time returned has a precision of 1 second.
@@ -3768,6 +3784,18 @@ void init_psutil_windows(void)
         INITERROR;
     }
 
+    // Exceptions.
+    TimeoutExpired = PyErr_NewException(
+        "_psutil_windows.TimeoutExpired", NULL, NULL);
+    Py_INCREF(TimeoutExpired);
+    PyModule_AddObject(module, "TimeoutExpired", TimeoutExpired);
+
+    TimeoutAbandoned = PyErr_NewException(
+        "_psutil_windows.TimeoutAbandoned", NULL, NULL);
+    Py_INCREF(TimeoutAbandoned);
+    PyModule_AddObject(module, "TimeoutAbandoned", TimeoutAbandoned);
+
+    // version constant
     PyModule_AddIntConstant(module, "version", PSUTIL_VERSION);
 
     // process status constants